reorganize and clean up intro page files a bit
