+++ src/applications/edu-benefits/0803/constants.js
@@ -2,3 +2,6 @@ export const TITLE =
   'Request for reimbursement of licensing or certification test fees';
 export const SUBTITLE =
   'Request for reimbursement of licensing or certification test fees (VA Form 22-0803)';
+export const OMB_RES_BURDEN = 15;
+export const OMB_NUMBER = '2900-0695';
+export const OMB_EXP_DATE = '01/31/2028';
