+++ src/applications/edu-benefits/0803/containers/App.jsx
@@ -6,7 +6,9 @@ import { isProfileLoading, isLoggedIn } from 'platform/user/selectors';
 import { useSelector } from 'react-redux';
 import formConfig from '../config/form';
 import manifest from '../manifest.json';
+
 import Breadcrumbs from '../components/Breadcrumbs';
+import NeedHelp from '../components/NeedHelp';
 
 export default function App({ location, children }) {
   const userLoggedIn = useSelector(state => isLoggedIn(state));
@@ -26,22 +28,14 @@ export default function App({ location, children }) {
   );
 
   return (
-    <div className="form-22-10297-container row">
+    <div className="form-22-0803-container row">
       <div className="vads-u-padding-left--0">
         <Breadcrumbs />
       </div>
       <RoutedSavableApp formConfig={formConfig} currentLocation={location}>
         {children}
       </RoutedSavableApp>
-      <va-need-help>
-        <div slot="content">
-          <p>
-            If you need help in completing this form, call VA TOLL-FREE at{' '}
-            <va-telephone contact="8884424551" vanity="888-GIBILL-1" />. If you
-            have hearing loss, call <va-telephone contact="711" tty />.
-          </p>
-        </div>
-      </va-need-help>
+      <NeedHelp />
     </div>
   );
 }
